+++ dmr/endpoint.py
@@ -279,7 +279,7 @@ class Endpoint:  # noqa: WPS214
             # And the last option is to handle error globally:
             return self._global_error_handler(controller, exc)
 
-    def get_operation(
+    def get_schema(
         self,
         path: str,
         pattern: URLPattern,